requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tableauhyperapi>=0.0.18000
dask>=2023.1.0
